task3/requirements.txt
Flask==2.3.2
gunicorn
mysql-connector-python
